requirements.txt
beautifulsoup4==4.12.2
numpy==1.26.2
pandas==2.1.3
protobuf==4.25.1
pypdf==3.17.1
Requests==2.31.0
streamlit==1.28.2
google-generativeai
textwrap3
streamlit_extras
extra_streamlit_components
st_pages


